refactor: simplify use live trains subscription
