+++ site/src/lib/digitraffic/hooks/use_live_trains_subscription.ts
@@ -1,41 +1,16 @@
 import type { SimplifiedTrain } from '@typings/simplified_train'
 import type { LocalizedStation } from '@lib/digitraffic'
 import type { Dispatch, SetStateAction } from 'react'
-import type { Train } from '@junat/digitraffic/types'
 
 import { useEffect, useState } from 'react'
 
-import { simplifyTrain } from '@utils/train'
+import { getNewTrains, trainsInFuture } from '@utils/train'
 import { useQuery } from '@tanstack/react-query'
-
-const getNewTrains = (
-  trains: SimplifiedTrain[],
-  updatedTrain: Train,
-  stationShortCode: string,
-  stations: LocalizedStation[],
-  type: 'ARRIVAL' | 'DEPARTURE' = 'DEPARTURE'
-) => {
-  return trains.map(train => {
-    if (
-      train.trainNumber === updatedTrain.trainNumber &&
-      train.scheduledTime ===
-        updatedTrain.timeTableRows.find(
-          tr => tr.stationShortCode === stationShortCode && tr.type === type
-        )?.scheduledTime
-    ) {
-      const t = simplifyTrain(updatedTrain, stationShortCode, stations)
-
-      return { t, ...train }
-    }
-
-    return train
-  })
-}
 
 interface UseLiveTrainsSubscriptionProps {
   stationShortCode: string
   type?: 'DEPARTURE' | 'ARRIVAL'
-  stations: LocalizedStation[] | undefined
+  stations: LocalizedStation[]
   initialTrains: SimplifiedTrain[]
 }
 
@@ -52,22 +27,19 @@ export const useLiveTrainsSubscription = ({
 ] => {
   const [trains, setTrains] = useState<SimplifiedTrain[]>(initialTrains)
 
-  const clientQuery = useQuery([LIVE_TRAINS_CLIENT_QUERY_KEY], async () => {
-    const { subscribeToStation } = await import('@junat/digitraffic-mqtt')
+  const { data: client } = useQuery(
+    [LIVE_TRAINS_CLIENT_QUERY_KEY],
+    async () => {
+      const { subscribeToStation } = await import('@junat/digitraffic-mqtt')
 
-    return subscribeToStation(stationShortCode)
-  })
+      return await subscribeToStation(stationShortCode)
+    }
+  )
 
   useEffect(() => {
-    if (
-      clientQuery.isFetching ||
-      !clientQuery.data ||
-      !stations ||
-      stations.length === 0
-    ) {
+    if (!client) {
       return
     }
-    const client = clientQuery.data
 
     ;(async () => {
       for await (const updatedTrain of client.trains) {
@@ -88,12 +60,7 @@ export const useLiveTrainsSubscription = ({
             type
           )
 
-          return newTrains.filter(train => {
-            return (
-              Date.parse(train.liveEstimateTime || train.scheduledTime) >
-              Date.now()
-            )
-          })
+          return trainsInFuture(newTrains)
         })
       }
     })()
@@ -102,13 +69,7 @@ export const useLiveTrainsSubscription = ({
       client.close()
       client.trains.return()
     }
-  }, [
-    clientQuery.data,
-    clientQuery.isFetching,
-    stationShortCode,
-    stations,
-    type
-  ])
+  }, [client, stationShortCode, stations, type])
 
   return [trains, setTrains]
 }
